Make a separate query to evaluate disabled and invisible setting status
