+++ apps/meteor/client/views/admin/settings/Setting/Setting.tsx
@@ -10,7 +10,7 @@ import { useTranslation } from 'react-i18next';
 
 import MemoizedSetting from './MemoizedSetting';
 import MarkdownText from '../../../../components/MarkdownText';
-import { useEditableSetting, useEditableSettingsDispatch } from '../../EditableSettingsContext';
+import { useEditableSetting, useEditableSettingsDispatch, useEditableSettingVisibilityQuery } from '../../EditableSettingsContext';
 import { useHasSettingModule } from '../hooks/useHasSettingModule';
 
 type SettingProps = {
@@ -96,7 +96,10 @@ function Setting({ className = undefined, settingId, sectionChanged }: SettingPr
 		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, [setting.value, (setting as ISettingColor).editor, update, persistedSetting]);
 
-	const { _id, disabled, readonly, type, packageValue, i18nLabel, i18nDescription, alert, invisible } = setting;
+	const { _id, readonly, type, packageValue, i18nLabel, i18nDescription, alert } = setting;
+
+	const disabled = !useEditableSettingVisibilityQuery(persistedSetting.enableQuery);
+	const invisible = !useEditableSettingVisibilityQuery(persistedSetting.displayQuery);
 
 	const labelText = (i18n.exists(i18nLabel) && t(i18nLabel)) || (i18n.exists(_id) && t(_id)) || i18nLabel || _id;
 
@@ -162,7 +165,7 @@ function Setting({ className = undefined, settingId, sectionChanged }: SettingPr
 			showUpgradeButton={showUpgradeButton}
 			sectionChanged={sectionChanged}
 			{...setting}
-			disabled={setting.disabled || shouldDisableEnterprise}
+			disabled={disabled || shouldDisableEnterprise}
 			value={value}
 			editor={editor}
 			hasResetButton={hasResetButton}
